pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Task manager"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dj-database-url>=2.3.0",
    "django>=5.1.4",
    "django-bootstrap5>=24.3",
    "django-filter>=24.3",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "rollbar>=1.1.0",
    "whitenoise>=6.8.2",
]

[dependency-groups]
dev = [
    "coverage>=7.6.9",
    "ruff>=0.8.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["task_manager"]
